Add main server with config and error handling
The main server file handles the creation of routes and handling any
internal server errors. If an error occurs, the server will render the
error view to the user, detailing the error message (includes the
stacktrace if running in development mode).

The routes for actual site pages will be added in the future, through
the use of express routers.