Add gulp sass and initial homepage content
Sass is converted into CSS through the gulp sass command. This allows
for structure within the stylesheet, using partials and modules to
separate repeated code into different files.

Initial homepage content has been added, including a jumbotron,
instruction thumbnails and the page footer. A search bar will be added
within the jumbotron, with a navbar sitting on top.

The logo and more detailed homepage content are yet to be added.