Add fixed navbar to the top of the page
The navbar was originally going to be transparent and become more
opaque as the user scrolled down the page, but I found that the solid
theme color brown looks better and is more consistent overall with the
site.

More navbar items will need to be added when necessary, but for now
contains placeholder items.