Add routes for the explore page
The explore page displays all results matching a search query, or shows
a list of all the letters if there is no search query.

The index view has been moved to index/index so that each routes file
has its own views subfolder. This will help pad files out to help with
readability, especially when more views are added in the future.