Add search results page with placeholder items
The search results page currently has a section for search as well as
displaying any search results in a grid view.

Explore routes and views have been renamed to search, as clicking the
explore button leads to the search page with no results. In the future,
going to the search page without specifying a query will show a list
of search tips and tricks.