Add search functionality for exact title matches
Any title that is entered into the searchbar is returned as an XML
string if it exists in the BaseX database. If not, the results are null
and an error message is displayed.

This search functionality will need to be improved to be contains
instead of exact matching.

Queries are made using the node-basex package which connects to the
REST API for BaseX.