Add data to explore and add letter route
The data for the explore page is a list of the first 8 results from the
query to the BaseX database. Example pagination (no styling) has been
added below the results.

The letter route currently just loads the XML for the specified letter
and prints it to the console. This will be formatted and displayed on
the screen in the future.