Fix pagination and basic search
Pagination had an issue whereby it was only changing 1 result as the
page changed. This explains the previous issue where it seemded to have
8 times as many results as it should have.

Basic search now checks for case-insensitive queries. For this reason,
having no query specified in the /explore route runs teh basic search
with an empty query as it matches everything.