Add xquery search and raw output view
XQuery search allows for searching by writing xquery strings. These
strings must be prepended with "XQ " in order to be acknowledged by the
server, otherwise they will be treated as a basic search.

Raw view allows the user to see exactly what BaseX is giving back, and
is in a tab on the explore page in similar fashion to the letter's xml
view.